Hijack onclick for clicked elements
Where possible, prevents the default onclick() event from firing
when tapping on a clicked element. For instance, if the user
clicks on a hidden link, the link will reveal itself without
following the link. This allows the user to reveal hidden text
without side effects.

Note that this will unfortunately not work for elements wrapped
in an <a> tag (or other tag that triggers click events) that
contains no text, since while the tag containing the text will
be flagged, the tag which fires the event (its parent) will not.